chore(portfollio-borrow): move portfollio borrow logic within mvvm
